- Adds IPv6 support in the api-server and wireguard driver
- Details in #637 and the design doc
[dual-stack.md.md](https://github.com/nexodus-io/nexodus/blob/main/docs/development/design/dual-stack.md.md)

+++ internal/nexodus/route_windows.go
@@ -2,11 +2,18 @@
 
 package nexodus
 
-import "github.com/nexodus-io/nexodus/internal/models"
+import (
+	"github.com/nexodus-io/nexodus/internal/models"
+	"github.com/nexodus-io/nexodus/internal/util"
+)
 
 // handlePeerRoute when a new configuration is deployed, delete/add the peer allowedIPs
 func (ax *Nexodus) handlePeerRouteOS(wgPeerConfig wgPeerConfig) {
 	for _, allowedIP := range wgPeerConfig.AllowedIPs {
+		// if the host does not support v6, skip adding the route
+		if util.IsIPv6Prefix(allowedIP) && !ax.ipv6Supported {
+			continue
+		}
 		if err := AddRoute(allowedIP, ax.tunnelIface); err != nil {
 			ax.logger.Debugf("route add failed: %v", err)
 		}
@@ -17,6 +24,10 @@ func (ax *Nexodus) handlePeerRouteOS(wgPeerConfig wgPeerConfig) {
 func (ax *Nexodus) handlePeerRouteDeleteOS(dev string, wgPeerConfig models.Device) {
 	// TODO: Windoze route lookups
 	for _, allowedIP := range wgPeerConfig.AllowedIPs {
+		// if the host does not support v6, skip adding the route
+		if util.IsIPv6Prefix(allowedIP) && !ax.ipv6Supported {
+			continue
+		}
 		if err := DeleteRoute(allowedIP, dev); err != nil {
 			ax.logger.Debug(err)
 		}
